pyproject.toml
[project]
name = "zap-crm"
version = "0.1.0"
description = "WhatsApp CRM built with Django"
requires-python = ">=3.11"
dependencies = [
    "django==4.2.16",
    "djangorestframework==3.15.2",
    "channels==4.1.0",
    "channels-redis==4.2.0",
    "celery==5.4.0",
    "redis==5.1.1",
    "psycopg2-binary==2.9.9",
    "cryptography>=42.0.0",
    "python-decouple==3.8",
    "daphne==4.1.2",
    "requests==2.32.3",
]

[dependency-groups]
dev = [
    "pytest-django>=4.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "factory-boy>=3.3.0",
    "faker>=24.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
